Final app

+++ app/Http/Controllers/PostsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 use App\Post;
 use DB;
@@ -10,13 +11,22 @@ use DB;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
-        
         $posts = Post::all();
         //return Post::where('title', 'Post two')->get();
         //$posts = DB::select('SELECT * From posts');
@@ -46,14 +56,36 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body'  => 'required'
+            'body'  => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
 
+        // Handle file upload
+        if($request->hasFile('cover_image')){
+
+            // Get file name with ext
+            $fileNamewithExt = $request->file('cover_image')->getClientOriginalName();
+
+            // get just file name
+            $fileName = pathinfo($fileNamewithExt, PATHINFO_FILENAME);
+
+            // get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+
+            // file to store
+            $fileNameToStore = $fileName.'_'.time().'.'.$extension;
+
+            //upload file
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
         // create post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post create');
@@ -80,6 +112,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -97,10 +135,32 @@ class PostsController extends Controller
             'body'  => 'required'
         ]);
 
+        // Handle file upload
+        if($request->hasFile('cover_image')){
+
+            // Get file name with ext
+            $fileNamewithExt = $request->file('cover_image')->getClientOriginalName();
+
+            // get just file name
+            $fileName = pathinfo($fileNamewithExt, PATHINFO_FILENAME);
+
+            // get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+
+            // file to store
+            $fileNameToStore = $fileName.'_'.time().'.'.$extension;
+
+            //upload file
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
+
         // create post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Updated');
@@ -115,6 +175,17 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
+        if($post->cover_image != 'noimage.jpg'){
+            // Delete image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+        
         $post->delete();
 
         return redirect('/posts')->with('success', 'Post Removed');
